Refactor images and stackscripts
