+++ src/services/images/index.ts
@@ -1,0 +1,9 @@
+export {
+  getImages,
+  getImage,
+  getImagesPage,
+  getLinodeImages,
+  createImage,
+  updateImage,
+  deleteImage,
+} from './images';
