+++ src/services/stackscripts/index.ts
@@ -1,0 +1,7 @@
+export {
+  getStackscripts,
+  getStackScript,
+  createStackScript,
+  updateStackScript,
+  deleteStackScript
+} from './stackscripts';
